+++ CHANGES.rst
@@ -1,10 +1,17 @@
 CHANGES
 =======
 
+0.1.7 (2012.09.27)
+------------------
+
+* Added preventDefault() and blur() to summarySelecter on-click.
+* Added initialSlideSpeed, openCallback, and closeCallback options.
+
 0.1.6 (2011.11.17)
 ------------------
 
-* Clicking ignoredElements does not trigger expand/collapse only if they are within summarySelector element.
+* Clicking ignoredElements does not trigger expand/collapse only if they are
+  within summarySelector element.
 
 0.1.5 (2011.07.14)
 ------------------